Local storage key for state includes registration if available
State storing in xAPI must take into account the registration param.
To keep the state in cache really related to the state in the LRS,
the key used for the state in localstorage must incorporate the
registration too. Otherwise, the same state (a generic 'xapi_state')
would be used/applied to launches with different registrations, which
is not correct.